accel/amdxdna: move coredump functions to amdxdna_coredump.c
Move PCI coredump handling into a dedicated source and header so its
callers depend on an explicit interface while AUX builds remain unchanged.

(cherry picked from commit 6c3789f1437c8b3a59f92c0cf7d57d495fb82cdd)
