+++ drivers/accel/amdxdna/Kbuild
@@ -56,6 +56,7 @@ amdxdna-$(OFT_CONFIG_AMDXDNA_PCI) += \
 	aie4_pci.o \
 	amdxdna_dpt.o \
 	amdxdna_error.o \
+	amdxdna_coredump.o \
 	npu1_regs.o \
 	npu3_regs.o \
 	npu4_regs.o \
